+++ closet-infra/src/main/java/project/closet/api/response/Header.java
@@ -1,4 +1,4 @@
-package project.closet.batch.api.response;
+package project.closet.api.response;
 
 public record Header(String resultCode, String resultMsg) {
