+++ closet-infra/src/main/java/project/closet/api/response/KakaoAddressResponse.java
@@ -1,4 +1,4 @@
-package project.closet.batch.api.response;
+package project.closet.api.response;
 
 import java.util.List;
 import lombok.Getter;
