+++ closet-infra/src/main/java/project/closet/api/response/Response.java
@@ -1,4 +1,4 @@
-package project.closet.batch.api.response;
+package project.closet.api.response;
 
 public record Response(Header header, Body body) {
