+++ closet-infra/src/main/java/project/closet/api/response/WeatherApiResponse.java
@@ -1,4 +1,4 @@
-package project.closet.batch.api.response;
+package project.closet.api.response;
 
 public record WeatherApiResponse(Response response) {
